refactor(studio): consolidate to single dev-studio system
Deleted parallel implementations:
- apps/admin/app/admin/editor/ (duplicate of dev-studio, used components/editor/*)
- components/editor/ (CodeEditor, XTerminal, FileTree, AIChat — all duplicated
  in components/dev-studio/ with inferior implementations)
- components/dev-studio/Terminal.tsx (fake static ANSI output, never a real shell)
- lib/devstudio/webcontainer/ (WebContainerRuntime built but never instantiated)
- _archived/api-studio/ (17 archived routes, fully superseded by /api/devstudio/*)
- _archived/store/ai-studio/ (archived page stub)
- Editor nav entry removed from AdminNav

Upgraded dev-studio components with better implementations from components/editor/:
- CodeEditor: lazy Monaco (dynamic import) replaces eager import
- XTerminal: real xterm.js with onCommand callback replaces fake ANSI art
- FileTree: takes string[] (flat paths) matching DevStudioClient state shape

Wired XTerminal.onCommand to POST /api/devstudio/execute with SSE streaming.
Workflow dispatch buttons (Deploy LMS, Deploy Admin, CI, Lint) remain in the
Terminal tab toolbar above the terminal — status lines shown in button strip.

Single canonical path: /admin/dev-studio → components/dev-studio/* → /api/devstudio/*
