--- _archived/api-studio/comments/route.ts
@@ -1,146 +1,0 @@
-// PUBLIC ROUTE: studio comments — auth inside handler
-import { getAdminClient } from '@/lib/supabase/admin';
-
-import { NextRequest, NextResponse } from 'next/server';
-
-import { applyRateLimit } from '@/lib/api/withRateLimit';
-import { withApiAudit } from '@/lib/audit/withApiAudit';
-export const runtime = 'nodejs';
-
-export const dynamic = 'force-dynamic';
-
-async function _GET(req: NextRequest) {
-  
-    const rateLimited = await applyRateLimit(req, 'api');
-    if (rateLimited) return rateLimited;
-const userId = req.headers.get('x-user-id');
-  const repoId = req.nextUrl.searchParams.get('repo_id');
-  const filePath = req.nextUrl.searchParams.get('file_path');
-  const branch = req.nextUrl.searchParams.get('branch');
-  
-  if (!userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
-  const supabase = await getAdminClient();
-  
-  let query = supabase
-    .from('studio_comments')
-    .select('*')
-    .eq('user_id', userId)
-    .order('line_start', { ascending: true });
-    
-  if (repoId) query = query.eq('repo_id', repoId);
-  if (filePath) query = query.eq('file_path', filePath);
-  if (branch) query = query.eq('branch', branch);
-
-  const { data, error } = await query;
-
-  if (error) {
-    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
-  }
-
-  return NextResponse.json(data || []);
-}
-
-async function _POST(req: NextRequest) {
-    const rateLimited = await applyRateLimit(req, 'api');
-    if (rateLimited) return rateLimited;
-
-  const userId = req.headers.get('x-user-id');
-  if (!userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
-  const { repo_id, file_path, branch, line_start, line_end, content } = await req.json();
-  if (!repo_id || !file_path || line_start === undefined || !content) {
-    return NextResponse.json({ error: 'Missing required fields' }, { status: 400 });
-  }
-
-  const supabase = await getAdminClient();
-  const { data, error } = await supabase
-    .from('studio_comments')
-    .insert({
-      user_id: userId,
-      repo_id,
-      file_path,
-      branch,
-      line_start,
-      line_end,
-      content
-    })
-    .select()
-    .maybeSingle();
-
-  if (error) {
-    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
-  }
-
-  return NextResponse.json(data);
-}
-
-async function _PUT(req: NextRequest) {
-  
-    const rateLimited = await applyRateLimit(req, 'api');
-    if (rateLimited) return rateLimited;
-const userId = req.headers.get('x-user-id');
-  if (!userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
-  const { id, content, resolved } = await req.json();
-  if (!id) {
-    return NextResponse.json({ error: 'id required' }, { status: 400 });
-  }
-
-  const supabase = await getAdminClient();
-  const updates: Record<string, unknown> = { updated_at: new Date().toISOString() };
-  if (content !== undefined) updates.content = content;
-  if (resolved !== undefined) updates.resolved = resolved;
-
-  const { data, error } = await supabase
-    .from('studio_comments')
-    .update(updates)
-    .eq('id', id)
-    .eq('user_id', userId)
-    .select()
-    .single();
-
-  if (error) {
-    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
-  }
-
-  return NextResponse.json(data);
-}
-
-async function _DELETE(req: NextRequest) {
-  
-    const rateLimited = await applyRateLimit(req, 'api');
-    if (rateLimited) return rateLimited;
-const userId = req.headers.get('x-user-id');
-  if (!userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
-  const { id } = await req.json();
-  if (!id) {
-    return NextResponse.json({ error: 'id required' }, { status: 400 });
-  }
-
-  const supabase = await getAdminClient();
-  const { error } = await supabase
-    .from('studio_comments')
-    .delete()
-    .eq('id', id)
-    .eq('user_id', userId);
-
-  if (error) {
-    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
-  }
-
-  return NextResponse.json({ ok: true });
-}
-export const GET = withApiAudit('/api/studio/comments', _GET);
-export const POST = withApiAudit('/api/studio/comments', _POST);
-export const PUT = withApiAudit('/api/studio/comments', _PUT);
-export const DELETE = withApiAudit('/api/studio/comments', _DELETE);
