--- _archived/api-studio/settings/route.ts
@@ -1,71 +1,0 @@
-// PUBLIC ROUTE: studio settings — auth inside handler
-import { getAdminClient } from '@/lib/supabase/admin';
-
-import { NextRequest, NextResponse } from 'next/server';
-
-import { applyRateLimit } from '@/lib/api/withRateLimit';
-import { withApiAudit } from '@/lib/audit/withApiAudit';
-export const runtime = 'nodejs';
-
-export const dynamic = 'force-dynamic';
-
-async function _GET(req: NextRequest) {
-  
-    const rateLimited = await applyRateLimit(req, 'api');
-    if (rateLimited) return rateLimited;
-const userId = req.headers.get('x-user-id');
-  if (!userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
-  const supabase = await getAdminClient();
-  const { data, error } = await supabase
-    .from('studio_settings')
-    .select('*')
-    .eq('user_id', userId)
-    .maybeSingle();
-
-  if (error && error.code !== 'PGRST116') {
-    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
-  }
-
-  return NextResponse.json(data || {
-    theme: 'dark',
-    font_size: 14,
-    word_wrap: true,
-    minimap: false,
-    auto_save: false,
-    keyboard_shortcuts: {}
-  });
-}
-
-async function _PUT(req: NextRequest) {
-  
-    const rateLimited = await applyRateLimit(req, 'api');
-    if (rateLimited) return rateLimited;
-const userId = req.headers.get('x-user-id');
-  if (!userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-
-  const body = await req.json();
-  const supabase = await getAdminClient();
-
-  const { data, error } = await supabase
-    .from('studio_settings')
-    .upsert({
-      user_id: userId,
-      ...body,
-      updated_at: new Date().toISOString()
-    }, { onConflict: 'user_id' })
-    .select()
-    .maybeSingle();
-
-  if (error) {
-    return NextResponse.json({ error: 'Internal server error' }, { status: 500 });
-  }
-
-  return NextResponse.json(data);
-}
-export const GET = withApiAudit('/api/studio/settings', _GET);
-export const PUT = withApiAudit('/api/studio/settings', _PUT);
