--- components/editor/FileTree.tsx
@@ -1,86 +1,0 @@
-'use client';
-
-import React from 'react';
-
-import { useState } from 'react';
-import { ChevronRight, ChevronDown, File, Folder } from 'lucide-react';
-
-interface FileNode {
-  name: string;
-  path: string;
-  type: 'file' | 'directory';
-  children?: FileNode[];
-}
-
-interface FileTreeProps {
-  files: FileNode[];
-  onFileSelect: (path: string) => void;
-  selectedFile?: string;
-}
-
-export default function FileTree({ files, onFileSelect, selectedFile }: FileTreeProps) {
-  const [expanded, setExpanded] = useState<Set<string>>(new Set(['/']));
-
-  const toggleExpand = (path: string) => {
-    const newExpanded = new Set(expanded);
-    if (newExpanded.has(path)) {
-      newExpanded.delete(path);
-    } else {
-      newExpanded.add(path);
-    }
-    setExpanded(newExpanded);
-  };
-
-  const renderNode = (node: FileNode, level: number = 0) => {
-    const isExpanded = expanded.has(node.path);
-    const isSelected = selectedFile === node.path;
-
-    return (
-      <div key={node.path}>
-        <div
-          className={`flex items-center gap-2 px-2 py-2 cursor-pointer hover:bg-slate-100 ${
-            isSelected ? 'bg-brand-blue-50 text-brand-blue-700' : ''
-          }`}
-          style={{ paddingLeft: `${level * 16 + 8}px` }}
-          onClick={() => {
-            if (node.type === 'directory') {
-              toggleExpand(node.path);
-            } else {
-              onFileSelect(node.path);
-            }
-          }}
-        >
-          {node.type === 'directory' ? (
-            <>
-              {isExpanded ? (
-                <ChevronDown className="w-4 h-4" />
-              ) : (
-                <ChevronRight className="w-4 h-4" />
-              )}
-              <Folder className="w-4 h-4 text-brand-blue-500" />
-            </>
-          ) : (
-            <>
-              <div className="w-4" />
-              <File className="w-4 h-4 text-slate-700" />
-            </>
-          )}
-          <span className="text-sm">{node.name}</span>
-        </div>
-
-        {node.type === 'directory' && isExpanded && node.children && (
-          <div>{node.children.map((child) => renderNode(child, level + 1))}</div>
-        )}
-      </div>
-    );
-  };
-
-  return (
-    <div className="h-full overflow-y-auto bg-white border-r">
-      <div className="p-2 border-b bg-slate-50">
-        <h3 className="font-semibold text-sm">Files</h3>
-      </div>
-      <div className="py-2">{files.map((file) => renderNode(file))}</div>
-    </div>
-  );
-}
